+++ src/DuetAPI/ObjectModel/Network/Network.cs
@@ -21,6 +21,17 @@
         public const string DefaultPassword = "reprap";
 
         /// <summary>
+        /// If this is set, the web server will allow cross-origin requests via the Access-Control-Allow-Origin header
+        /// </summary>
+        [LinuxProperty]
+        public string CorsSite
+        {
+            get =>_corsSite;
+            set => SetPropertyValue(ref _corsSite, value);
+        }
+        private string _corsSite;
+
+        /// <summary>
         /// Hostname of the machine
         /// </summary>
         public string Hostname