+++ src/DuetControlServer/Files/CodeFile.cs
@@ -3,7 +3,6 @@ using DuetAPI.Commands;
 using DuetAPI.ObjectModel;
 using Nito.AsyncEx;
 using System;
-using System.Buffers;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;