+++ src/DuetControlServer/Model/Updater.cs
@@ -3,7 +3,6 @@ using DuetAPI.ObjectModel;
 using Nito.AsyncEx;
 using System;
 using System.Collections.Concurrent;
-using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Text.Json;