+++ src/DuetPluginService/Program.cs
@@ -1,0 +1,12 @@
+﻿using System;
+
+namespace DuetPluginService
+{
+    class Program
+    {
+        static void Main(string[] args)
+        {
+            Console.WriteLine("Hello World!");
+        }
+    }
+}